+++ actfw_core/libcamera_capture.py
@@ -94,7 +94,13 @@ class LibcameraCapture(Producer[Frame[bytes]]):
     _requests: Optional[List[libcam.Request]]
     _camera_config: libcam.CameraConfiguration
 
-    def __init__(self, size: Tuple[int, int], pixel_format: libcam.PixelFormat, camera_index: int = 0) -> None:
+    def __init__(
+        self,
+        size: Tuple[int, int],
+        pixel_format: libcam.PixelFormat,
+        camera_index: int = 0,
+        orientation: libcam.Orientation = libcam.Orientation.Rotate0,
+    ) -> None:
         """
         NOTE: BGR を指定すると実際には RGB で取得される
         """
@@ -108,6 +114,7 @@ class LibcameraCapture(Producer[Frame[bytes]]):
         self._camera = self._cm.cameras[camera_index]
         self._camera.acquire()
         self._camera_config = self._camera.generate_configuration([libcam.StreamRole.Viewfinder])
+        self._camera_config.orientation = orientation
         stream_config: libcam.StreamConfiguration = self._camera_config.at(0)
         stream_config.size = libcam.Size(*self._size)
         stream_config.pixel_format = self._pixel_format